Move lib/*Patch.rb to lib/patches/
