+++ lib/patches/rack_oauth2_client_patch.rb
